--- src/app/components/Link/constants.js
@@ -1,2 +1,0 @@
-
-export const DOMAIN_NAME = "analog.cafe"
