--- src/app/components/List/constants.js
@@ -1,8 +1,0 @@
-// tools
-
-// components
-
-// state
-
-// dictionary
-export const ROUTE_ARTICLE_DIR = "/zine"
