--- src/app/containers/Composer/containers/ContentEditor/state.json
@@ -1,10 +1,0 @@
-{
-  "nodes":[
-    {
-      "data":{ },
-      "kind":"block",
-      "isVoid":false,
-      "type":"paragraph"
-    }
-  ]
-}
