--- src/app/containers/Composer/containers/HeaderEditor/state.json
@@ -1,1 +1,0 @@
-{ "title": "", "subtitle": "" }
