--- src/app/containers/Picture/constants.js
@@ -1,8 +1,0 @@
-// tools
-
-// components
-
-// state
-
-// dictionary
-export const ROUTE_IMAGE_API = "/api/image"
