--- src/app/containers/Post/constants.js
@@ -1,9 +1,0 @@
-// tools
-
-// components
-
-// state
-
-// dictionary
-export const ROUTE_POST_API = "/api/post"
-export const ROUTE_ARTICLE_DIR = "/zine"
