+++ src/constants/composer.js
@@ -1,0 +1,11 @@
+export const DEFAULT_COMPOSER_HEADER_STATE = { "title": "", "subtitle": "" }
+export const DEFAULT_COMPOSER_EDITOR_STATE = {
+  "nodes":[
+    {
+      "data":{ },
+      "kind":"block",
+      "isVoid":false,
+      "type":"paragraph"
+    }
+  ]
+}
