+++ src/constants/link.js
@@ -1,0 +1,1 @@
+export const DOMAIN_NAME = "analog.cafe"
