+++ src/constants/picture.js
@@ -1,0 +1,1 @@
+export const ROUTE_IMAGE_API = "/api/image"
