+++ src/constants/post.js
@@ -1,0 +1,2 @@
+export const ROUTE_POST_API = "/api/post"
+export const ROUTE_ARTICLE_DIR = "/zine"
